requirements.txt
langchain>=0.0.267
langchain-community>=0.0.6
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
elasticsearch>=8.9.0
pymilvus>=2.3.0
openai>=1.0.0
numpy>=1.25.2
tiktoken>=0.5.1
pandas>=2.0.0
python-dotenv>=1.0.0
pypdf>=3.15.1
docx2txt>=0.8
unstructured>=0.10.0
nltk>=3.8.1
scikit-learn>=1.3.0
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
tqdm>=4.66.1
sentence-transformers>=2.2.2 